Move BlockMaster to block package
